+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased](https://github.com/openfga/go-sdk/compare/v0.7.2...HEAD)
 
+- feat: add support for custom headers per request
 - feat: add generic `ToPtr[T any](v T) *T` function for creating pointers to any type
 - deprecation: `PtrBool`, `PtrInt`, `PtrInt32`, `PtrInt64`, `PtrFloat32`, `PtrFloat64`, `PtrString`, and `PtrTime` are now deprecated in favor of the generic `ToPtr` function